More octree into MeshGeometryModel3D. This is more proper than put it in GeometryModel3D
