+++ Source/HelixToolkit.Wpf.SharpDX/Model/Elements3D/AbstractElements3D/GeometryModel3D.cs
@@ -27,7 +27,6 @@ namespace HelixToolkit.Wpf.SharpDX
     public abstract class GeometryModel3D : Model3D, IHitable, IBoundable, IVisible, IThrowingShadow, ISelectable, IMouse3D
     {
         protected RasterizerState rasterState;
-        protected GeometryOctree octree;
         /// <summary>
         /// Override in derived classes to specify the
         /// size, in bytes, of the vertices used for rendering.
@@ -65,34 +64,6 @@ namespace HelixToolkit.Wpf.SharpDX
             }
         }
 
-        public static readonly DependencyProperty UseOctreeHitTestProperty = DependencyProperty.Register("UseOctreeHitTest", typeof(bool), typeof(GeometryModel3D), new PropertyMetadata(true, 
-            (s,e)=> 
-            {
-                var m = s as GeometryModel3D;
-                if((bool)e.NewValue == true)
-                {
-                    m.UpdateOctree();
-                }
-                else
-                {
-                    m.octree = null;
-                }
-            }));
-        /// <summary>
-        /// Use octree for hittest instead of bruteforce hit test for better performance
-        /// </summary>
-        public bool UseOctreeHitTest
-        {
-            set
-            {
-                SetValue(UseOctreeHitTestProperty, value);
-            }
-            get
-            {
-                return (bool)GetValue(UseOctreeHitTestProperty);
-            }
-        }
-
         public static readonly DependencyProperty GeometryProperty =
             DependencyProperty.Register("Geometry", typeof(Geometry3D), typeof(GeometryModel3D), new UIPropertyMetadata(GeometryChanged));
 
@@ -127,33 +98,11 @@ namespace HelixToolkit.Wpf.SharpDX
             //b.Maximum = Vector3.TransformCoordinate(b.Maximum, m);
             this.Bounds = b;
             //this.BoundsDiameter = (b.Maximum - b.Minimum).Length();
-            UpdateOctree();
             if (renderHost !=null)
             {
                 var host = this.renderHost;
                 this.Detach();
                 this.Attach(host);
-            }
-        }
-
-        protected void UpdateOctree()
-        {
-            if(UseOctreeHitTest && IsHitTestVisible && Geometry != null && Geometry.Positions != null 
-                && Geometry.Indices != null && Geometry.Positions.Count > 0 && Geometry.Indices.Count > 0)
-            {
-                this.octree = new GeometryOctree(this.Geometry.Positions, this.Geometry.Indices);
-#if DEBUG
-                var sw = Stopwatch.StartNew();
-#endif
-                this.octree.UpdateTree();
-#if DEBUG
-                sw.Stop();
-                Debug.WriteLine("Buildtree time =" + sw.ElapsedMilliseconds);
-#endif
-            }
-            else
-            {
-                this.octree = null;
             }
         }
 
@@ -409,78 +358,57 @@ namespace HelixToolkit.Wpf.SharpDX
             {
                 return false;
             }
-#if DEBUG
-            Stopwatch sw = Stopwatch.StartNew();
-#endif
+
+            var g = this.Geometry as MeshGeometry3D;
             bool isHit = false;
-            if (UseOctreeHitTest && octree!=null)
-            {
-                isHit = octree.HitTest(this, modelMatrix, rayWS, ref hits);
-            }
-            else
-            {
-                var g = this.Geometry as MeshGeometry3D;
-                isHit = false;
-                var result = new HitTestResult();
-                result.Distance = double.MaxValue;
+            var result = new HitTestResult();
+            result.Distance = double.MaxValue;
 
-                if (g != null)
+            if (g != null)
+            {
+                var m = this.modelMatrix;
+
+                // put bounds to world space
+                var b = BoundingBox.FromPoints(this.Bounds.GetCorners().Select(x => Vector3.TransformCoordinate(x, m)).ToArray());
+
+                //var b = this.Bounds;
+
+                // this all happens now in world space now:
+                if (rayWS.Intersects(ref b))
                 {
-                    var m = this.modelMatrix;
-
-                    // put bounds to world space
-                    var b = BoundingBox.FromPoints(this.Bounds.GetCorners().Select(x => Vector3.TransformCoordinate(x, m)).ToArray());
-
-                    //var b = this.Bounds;
-
-                    // this all happens now in world space now:
-                    if (rayWS.Intersects(ref b))
+                    int index = 0;
+                    foreach (var t in g.Triangles)
                     {
-                        int index = 0;
-                        foreach (var t in g.Triangles)
+                        float d;
+                        var p0 = Vector3.TransformCoordinate(t.P0, m);
+                        var p1 = Vector3.TransformCoordinate(t.P1, m);
+                        var p2 = Vector3.TransformCoordinate(t.P2, m);
+                        if (Collision.RayIntersectsTriangle(ref rayWS, ref p0, ref p1, ref p2, out d))
                         {
-                            float d;
-                            var p0 = Vector3.TransformCoordinate(t.P0, m);
-                            var p1 = Vector3.TransformCoordinate(t.P1, m);
-                            var p2 = Vector3.TransformCoordinate(t.P2, m);
-                            if (Collision.RayIntersectsTriangle(ref rayWS, ref p0, ref p1, ref p2, out d))
+                            if (d > 0 && d < result.Distance) // If d is NaN, the condition is false.
                             {
-                                if (d > 0 && d < result.Distance) // If d is NaN, the condition is false.
-                                {
-                                    result.IsValid = true;
-                                    result.ModelHit = this;
-                                    // transform hit-info to world space now:
-                                    result.PointHit = (rayWS.Position + (rayWS.Direction * d)).ToPoint3D();
-                                    result.Distance = d;
+                                result.IsValid = true;
+                                result.ModelHit = this;
+                                // transform hit-info to world space now:
+                                result.PointHit = (rayWS.Position + (rayWS.Direction * d)).ToPoint3D();
+                                result.Distance = d;
 
-                                    var n = Vector3.Cross(p1 - p0, p2 - p0);
-                                    n.Normalize();
-                                    // transform hit-info to world space now:
-                                    result.NormalAtHit = n.ToVector3D();// Vector3.TransformNormal(n, m).ToVector3D();
-                                    result.TriangleIndices = new System.Tuple<int, int, int>(g.Indices[index], g.Indices[index + 1], g.Indices[index + 2]);
-                                    isHit = true;
-                                }
+                                var n = Vector3.Cross(p1 - p0, p2 - p0);
+                                n.Normalize();
+                                // transform hit-info to world space now:
+                                result.NormalAtHit = n.ToVector3D();// Vector3.TransformNormal(n, m).ToVector3D();
+                                result.TriangleIndices = new System.Tuple<int, int, int>(g.Indices[index], g.Indices[index + 1], g.Indices[index + 2]);
+                                isHit = true;
                             }
-                            index += 3;
                         }
+                        index += 3;
                     }
                 }
-                if (isHit)
-                {
-                    hits.Add(result);
-                }
             }
-#if DEBUG
-            sw.Stop();
             if (isHit)
             {
-                Debug.WriteLine("Octree time:" + sw.ElapsedMilliseconds);
-                foreach (var hit in hits)
-                {
-                    Debug.WriteLine(hit.PointHit);
-                }
+                hits.Add(result);
             }
-#endif
             return isHit;
         }
